Basic Authority Protection, Better Descriptions, and Smell Buff
I've ensured Police Chief doesn't allow Kinfolk. I'll update this for future PRs adding new jobs that are similar in requiring Innocent Mortal Only.
The Giovanni block for Kinfolk is a obvious one as well, because Kinfolk is passed on through genetics, it's unlikely any Kinfolk have made it inside of that family.
I've increased the smell to four as it's still close but doesn't require a werewolf player to awkwardly walk up to someone and literally sniff their face.