Update baron.dm
Includes Baali, Caitiff and Banu Haqim as allowed Barons. Caitiff should be allowed to be Barons as it's the most feasible position of 'power' one could obtain, disallowing an actual Prince as I consider the one example of that, the Prince of Cairo, to be an exception, not an example. Baali could be Barons for sure and would as it helps them blend in to perform their evil acts. Banu I am unsure but since Banu are no longer whitelisted, it would be good to give them more jobs.